Updates tests structure
